Fix possible race condition in SqlTaskManager
There is a possible race condition between SqlTaskManager and
 CatalogPruneTask on workers. Before the catalogs for a Task are set it
is possible for a active catalogs from the prune task to arrive, fail to
be agumentted by the catalogs of the task, but then prune the catalogs
after they are set and loaded. This fix uses a ReadWriteLock to ensure
mutual exclusion of task catalog setting and task scanning/pruning.