Remove unnecessary logic from CachingJdbcClient
This logically reverts commit b25c90b44fc47daa3522111ccce2c9f89d76ce4b.
The class uses `EvictableCacheBuilder` and the
`io.trino.cache.EvictableCacheBuilder#cacheDisabled` considers the cache
to be disabled when TTL is 0 or cache size is 0, so no need to set cache
size to 0 if TTL is known to be 0.