Add missing check in PreAggregateCaseAggregations
In 69ac97ac72a954471420f6c6c707d279691eb142, a check allowing for
the result of the call to optimize an expression to be a non-constant
expression was lost:

    Object defaultValue = optimizeExpression(caseExpression.defaultValue().get(), context);
    if (defaultValue != null) {
      ...
    }

As a result, the optimization produced plans with references to invalid columns.

This change restores the check and cleans up the nested if/else structure to make it more readable.