Do not specify what is the implementation of a dependency
The previous approach was kind of defeating the reason of
using the guice. It should be handled in guice modules to
tell what should be injected where. Classes itself should not have
such knowledge.

Thanks to this, there is no need to use guice just to
inject a single dependency to LazyConnectionFactory constructor in
TestLazyConnectionFactory.