Remove deprecated SystemAccessControlFactory#create method
Just over a year seems to be a long-enough transition period. The way
this method was introduced forced the implementors of the interface to
implement it even if the other, tracing-aware, version was implemented
as well - but it was never called directly by the engine. This caused
moderate chaos and confusion, resulting in versions which just returned
`null`.